viewer/requirements.txt
requests==2.31.0
Pillow==10.2.0
inky==1.5.0
spidev==3.6
RPi.GPIO==0.7.1
smbus2==0.4.3
